+++ core/Objects/CmfPackage.cs
@@ -153,6 +153,12 @@ namespace Cmf.CLI.Core.Objects
         public bool? IsUniqueInstall { get; private set; }
 
         /// <summary>
+        /// Gets or sets whether package installation should be forced by Deployment Framework.
+        /// </summary>
+        [JsonProperty(Order = 16)]
+        public bool? IsToForceInstall { get; private set; }
+
+        /// <summary>
         /// Gets or sets the is root package.
         /// </summary>
         /// <value>
@@ -307,6 +313,7 @@ namespace Cmf.CLI.Core.Objects
         /// <param name="targetLayer">The target layer.</param>
         /// <param name="isInstallable">The is installable.</param>
         /// <param name="isUniqueInstall">The is unique install.</param>
+        /// <param name="isToForceInstall">The is to force install.</param>
         /// <param name="keywords">The keywords.</param>
         /// <param name="isToSetDefaultSteps">The is to set default steps.</param>
         /// <param name="dependencies">The dependencies.</param>
@@ -317,7 +324,7 @@ namespace Cmf.CLI.Core.Objects
         /// <param name="testPackages">The test Packages.</param>
         [JsonConstructor]
         public CmfPackage(string name, string packageId, string version, string description, PackageType packageType,
-                          string targetDirectory, string targetLayer, bool? isInstallable, bool? isUniqueInstall, string keywords,
+                  string targetDirectory, string targetLayer, bool? isInstallable, bool? isUniqueInstall, bool? isToForceInstall, string keywords,
                           bool? isToSetDefaultSteps, DependencyCollection dependencies, List<Step> steps,
                           List<ContentToPack> contentToPack, List<string> xmlInjection, bool? waitForIntegrationEntries, List<string> baseLocalizationFiles, DependencyCollection testPackages = null) : this()
         {
@@ -330,6 +337,7 @@ namespace Cmf.CLI.Core.Objects
             TargetLayer = targetLayer;
             IsInstallable = isInstallable ?? true;
             IsUniqueInstall = isUniqueInstall ?? false;
+            IsToForceInstall = isToForceInstall;
             Keywords = keywords;
             IsToSetDefaultSteps = isToSetDefaultSteps ?? true;
             Dependencies = dependencies;
@@ -431,6 +439,7 @@ namespace Cmf.CLI.Core.Objects
                    TargetLayer.IgnoreCaseEquals(other.TargetLayer) &&
                    IsInstallable == other.IsInstallable &&
                    IsUniqueInstall == other.IsUniqueInstall &&
+                   IsToForceInstall == other.IsToForceInstall &&
                    Keywords.IgnoreCaseEquals(other.Keywords) &&
                    XmlInjection.Equals(other.XmlInjection) &&
                    EqualityComparer<DependencyCollection>.Default.Equals(Dependencies, other.Dependencies) &&
@@ -882,6 +891,7 @@ namespace Cmf.CLI.Core.Objects
                 "",
                 false,
                 false,
+                tokens.ContainsKey("istoforceinstall") ? bool.Parse(tokens["istoforceinstall"]) : null,
                 tokens.ContainsKey("keywords") ? tokens["keywords"] : null,
                 true,
                 deps,